feat: re-wired all the project pages back together to work properly
We now have a project page area of the site which is wired in
correctly this time. It exists to document stuff that I am working
on at the time and will also never be updated. Probably want to work
on default pages for when I don't have anything to work on at the
time.

Also... 404s...